+++ lib/api/HttpServer.ts
@@ -11,7 +11,7 @@ import Zamza from "../Zamza";
 import { HttpConfig } from "../interfaces";
 import { routeRoot, routeTopicConfig, routeInfo,
          routeFetch, routeProduce, routeManage,
-         routeHook } from "./routes";
+         routeHook, routeReplay } from "./routes";
 
 import AccessControll from "./AccessControll";
 
@@ -54,8 +54,11 @@ export default class HttpServer {
         app.use(bodyParser.json());
 
         app.use("/", routeRoot(this.zamza));
+
         app.use("/api/config", routeTopicConfig(this.zamza));
         app.use("/api/config", routeHook(this.zamza));
+        app.use("/api/config", routeReplay(this.zamza));
+
         app.use("/api/info", routeInfo(this.zamza));
         app.use("/api/fetch", routeFetch(this.zamza));
         app.use("/api/produce", routeProduce(this.zamza));