+++ lib/api/routes/index.ts
@@ -5,3 +5,4 @@ export * from "./fetch";
 export * from "./produce";
 export * from "./manage";
 export * from "./hook";
+export * from "./replay";