+++ lib/db/models/index.ts
@@ -3,3 +3,4 @@ export * from "./TopicConfigModel";
 export * from "./TopicMetadata";
 export * from "./LockModel";
 export * from "./HookModel";
+export * from "./ReplayModel";