+++ lib/interfaces/Replay.ts
@@ -1,0 +1,5 @@
+export interface Replay {
+    topic: string;
+    timestamp: number;
+    consumerGroup: string;
+}